Use monaco-worker-manager
This saves some boilerplate related to dealing with web workers.
