+++ package.json
@@ -42,6 +42,7 @@
   "dependencies": {
     "@types/json-schema": "^7.0.0",
     "jsonc-parser": "^3.0.0",
+    "monaco-worker-manager": "^1.0.0",
     "path-browserify": "^1.0.0",
     "prettier": "2.0.5",
     "vscode-languageserver-textdocument": "^1.0.0",
